+++ storage/cache/adapter/Redis.php
@@ -9,6 +9,7 @@
 namespace lithium\storage\cache\adapter;
 
 use lithium\storage\Cache;
+use lithium\net\HostString;
 use Redis as RedisCore;
 
 /**
@@ -51,6 +52,11 @@ use Redis as RedisCore;
 class Redis extends \lithium\storage\cache\Adapter {
 
 	/**
+	 * The default host used to connect to the server.
+	 */
+	const DEFAULT_HOST = '127.0.0.1';
+
+	/**
 	 * The default port used to connect to the server.
 	 */
 	const DEFAULT_PORT = 6379;
@@ -75,8 +81,10 @@ class Redis extends \lithium\storage\cache\Adapter {
 	 *          is otherwise set. Can be either a `strtotime()` compatible tring or TTL in
 	 *          seconds. To indicate items should not expire use `Cache::PERSIST`. Defaults
 	 *          to `+1 hour`.
-	 *        - `'host'` _string_: A string in the form of `'host:port'` indicating the Redis server
-	 *          to connect to. Defaults to `'127.0.0.1:6379'`.
+	 *        - `'host'` _string_: A string in the form of `'<host>'`, `'<host>:<port>'` or
+	 *          `':<port>'` indicating the host and/or port to connect to. When one or both are
+	 *          not provided uses general server defaults.
+	 *          To use Unix sockets specify the path to the socket (i.e. `'/path/to/socket'`).
 	 *        - `'persistent'` _boolean_: Indicates whether the adapter should use a persistent
 	 *          connection when attempting to connect to the Redis server. If `true`, it will
 	 *          attempt to reuse an existing connection when connecting, and the connection will
@@ -85,9 +93,9 @@ class Redis extends \lithium\storage\cache\Adapter {
 	 */
 	public function __construct(array $config = array()) {
 		$defaults = array(
-			'host' => '127.0.0.1:6379',
 			'scope' => null,
 			'expiry' => '+1 hour',
+			'host' => static::DEFAULT_HOST . ':' . static::DEFAULT_PORT,
 			'persistent' => false
 		);
 		parent::__construct($config + $defaults);
@@ -103,12 +111,16 @@ class Redis extends \lithium\storage\cache\Adapter {
 		if (!$this->connection) {
 			$this->connection = new RedisCore();
 		}
-		list($address, $port) = $this->_formatHost($this->_config['host']);
 		$method = $this->_config['persistent'] ? 'pconnect' : 'connect';
-		if ($port) {
-			$this->connection->{$method}($address, $port);
+
+		if (HostString::isSocket($this->_config['host'])) {
+			$this->connection->{$method}($this->_config['host']);
 		} else {
-			$this->connection->{$method}($address);
+			$host = HostString::parse($this->_config['host']) + array(
+				'host' => static::DEFAULT_HOST,
+				'port' => static::DEFAULT_PORT
+			);
+			$this->connection->{$method}($host['host'], $host['port']);
 		}
 
 		if ($this->_config['scope']) {
@@ -154,22 +166,6 @@ class Redis extends \lithium\storage\cache\Adapter {
 			return true;
 		}
 		return is_callable(array($this->connection, $method));
-	}
-
-	/**
-	 * Formats standard `'host:port'` or `'/path/to/socket'` strings
-	 *
-	 * @param mixed $host A host string in `'host:port'` or `'/path/to/socket'` format
-	 * @return array Returns an array of `Redis` connection definitions.
-	 */
-	protected function _formatHost($host) {
-		if (strpos($host, ':') > 0) {
-			list($address, $port) = explode(':', $host);
-		} else {
-			$address = $host;
-			$port = strpos($host, '/') === 0 ? null : self::DEFAULT_PORT;
-		}
-		return array($address, $port);
 	}
 
 	/**